Starting a bit of code cleanup
